Fix the loss of ParserOutput pointer in ContentDOMTransformStages
When running a ContentDOMTransformStage, we effectively clone the input
ParserOutput, which is in contradiction with the current expectations of
the pipeline. This patch slightly modifies the logic by making it
possible to apply a PageBundle data to an existing ParserOutput without the
necessity to create a new one.

Bug: T364597
Change-Id: I633fc33485f22cf645acd41650a6983df3b0a534
(cherry picked from commit 2c0fe9319335d801673cc8cae23d19f00b230ee6)